Update proposed format to reflect POC
The POC ended up creating the bundle differently from what had
originally been written here. There was no data.tar.gz in the POC and it
also ended up needing an index file that included meta-data and a link
between the stored file and the URI referenced in the catalog.